# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servidor API para el Proyecto Sakoo (Arquitectura Limpia).",
        "title": "Sakoo Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte Sakoo",
            "url": "http://www.sakoo.com",
            "email": "soporte@sakoo.com"
        },
        "license": {
            "name": "Propietaria",
            "url": "http://www.sakoo.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/accounts/own": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite registrar una nueva cuenta bancaria propia (POST) o listar todas las cuentas propias asociadas al usuario autenticado (GET).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Propias"
                ],
                "summary": "Gestionar cuentas bancarias propias (Crear / Listar)",
                "parameters": [
                    {
                        "description": "Datos de la cuenta (requerido solo para POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.OwnAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error de autenticación o datos incorrectos",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite registrar una nueva cuenta bancaria propia (POST) o listar todas las cuentas propias asociadas al usuario autenticado (GET).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Propias"
                ],
                "summary": "Gestionar cuentas bancarias propias (Crear / Listar)",
                "parameters": [
                    {
                        "description": "Datos de la cuenta (requerido solo para POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.OwnAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error de autenticación o datos incorrectos",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/own/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite actualizar por completo una cuenta bancaria propia (PUT) o eliminarla de forma lógica (DELETE) especificando su ID en la ruta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Propias"
                ],
                "summary": "Gestionar detalle de cuenta propia (Actualizar / Eliminar)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la cuenta bancaria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la cuenta (requerido solo para PUT)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.OwnAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID inválido, no autorizado o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite actualizar por completo una cuenta bancaria propia (PUT) o eliminarla de forma lógica (DELETE) especificando su ID en la ruta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Propias"
                ],
                "summary": "Gestionar detalle de cuenta propia (Actualizar / Eliminar)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la cuenta bancaria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la cuenta (requerido solo para PUT)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.OwnAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID inválido, no autorizado o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/third-party": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite registrar una nueva cuenta bancaria de un tercero (POST) o listar todas las de terceros asociadas al usuario autenticado (GET).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Terceros"
                ],
                "summary": "Gestionar cuentas bancarias de terceros (Crear / Listar)",
                "parameters": [
                    {
                        "description": "Datos de la cuenta (requerido solo para POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ThirdPartyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite registrar una nueva cuenta bancaria de un tercero (POST) o listar todas las de terceros asociadas al usuario autenticado (GET).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Terceros"
                ],
                "summary": "Gestionar cuentas bancarias de terceros (Crear / Listar)",
                "parameters": [
                    {
                        "description": "Datos de la cuenta (requerido solo para POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ThirdPartyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/third-party/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite actualizar por completo una cuenta de terceros (PUT) o eliminarla físicamente (DELETE) especificando su ID en la ruta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Terceros"
                ],
                "summary": "Gestionar detalle de cuenta de terceros (Actualizar / Eliminar)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la cuenta de terceros",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la cuenta (requerido solo para PUT)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ThirdPartyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID inválido, no autorizado o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite actualizar por completo una cuenta de terceros (PUT) o eliminarla físicamente (DELETE) especificando su ID en la ruta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cuentas Bancarias - Terceros"
                ],
                "summary": "Gestionar detalle de cuenta de terceros (Actualizar / Eliminar)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la cuenta de terceros",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la cuenta (requerido solo para PUT)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ThirdPartyAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID inválido, no autorizado o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "track_code": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.OwnAccountRequest": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "holder_name": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ThirdPartyAccountRequest": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "alias": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "document_number": {
                    "type": "string"
                },
                "holder_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Coloque el token JWT de la siguiente manera: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}